use-cases/intelligent-negotiation-assistant-procurement/ui/requirements.txt
streamlit
requests
python-dotenv
plotly
pandas

